RedditAPITimeTrials/requirements.txt
praw~=7.4.0
pyfunctional~=1.4.3
psaw~=0.1.0